Added different options to initiate an errval
It is now possible to copy an errval 'a' by creating an 'b' as
follows: b = errval(a). To change the printout-option of this
copy close the printout with a !bang: printout='mynewopt!'.
Furthermore, tuples of length 2 (3) are accepted, whose entries
are then assigned to val and err (and printout).
Lastly, if the input doesn't fit any expected option a ValueError
is raised.